Refactor Const class to include payment modes and move getRandomDarkColor and getRandomLightColor to Utils
