+++ lib/const.dart
@@ -70,59 +70,11 @@ class Const {
     '2500variant': Colors.pink[900],
   };
 
-  final List<Color> darkColors = [
-    Colors.lightGreen,
-    Colors.redAccent,
-    Colors.blueAccent,
-    Colors.purpleAccent,
-    Colors.pinkAccent,
-    Colors.deepPurpleAccent,
-    Colors.lightBlueAccent,
-    Colors.indigoAccent,
-    Colors.brown,
-    Colors.blueGrey,
-    Colors.black,
-    Colors.grey,
-    Colors.deepOrange,
-    Colors.teal,
-    Colors.cyan,
-    Colors.orange,
-  ];
-  Color getRandomDarkColor() {
-    final random = Random();
-    return darkColors[random.nextInt(darkColors.length)];
-  }
-
-  final List<Color> lightColors = [
-    Colors.greenAccent,
-    Colors.lightBlueAccent,
-    Colors.lightGreenAccent,
-    Colors.amberAccent,
-    Colors.tealAccent,
-    const Color.fromARGB(255, 153, 249, 249),
-    Colors.limeAccent,
-    Color.fromARGB(255, 255, 169, 169),
-    const Color.fromARGB(255, 248, 139, 175),
-    Color.fromARGB(255, 235, 124, 255),
-  ];
-
-  Color getRandomLightColor() {
-    return lightColors[DateTime.now().millisecond % lightColors.length];
-  }
-
-  Future<String> getUserName() async {
-    var u = await LS().read('user_details');
-    if (u != null) {
-      var uu = jsonDecode(u);
-      UserDetails user = UserDetails.fromJson(uu);
-
-      if (user.name == null) {
-        return 'Username Error';
-      } else {
-        return user.name!;
-      }
-    } else {
-      return 'Username Error';
-    }
-  }
+  // payment modes
+  final paymentModes = {
+    'UPI': {'icon': "assets/images/icon_upi.png"},
+    'Cash': {'icon': "assets/images/icon_cash.png"},
+    'Card': {'icon': "assets/images/icon_card.png"},
+    'Gift': {'icon': "assets/images/icon_gratis.png"},
+  };
 }
